+++ src/components/DistributionSparkline/index.ts
@@ -1,0 +1,3 @@
+export * from "./DistributionSparkline";
+export * from "./variants";
+export * from "./domain";